2 Commits for 2 changes:
1. Factor references into their own parse rules and remove left-recursion
2. Add parser Listener that turns CST into AST on the fly and nulls out CST to save memory

### Performance Enhancement

Time required to parse and emit CHIRRTL

| design | master (.fir) | this PR (.fir) | .pb on both |
| --- | --- | --- | --- |
| small | 5.3 s | 5.3 s | 2.5 s |
| medium | 35 s | 33 s | 12.7 s |
| large | 170 s | 124 s | 41 s |

Note a huge gain for performance, but a wash or strict improvement. Still much slower than `.pb` parsing but could probably be helped by more (or any) parallelism.

### Memory Use Enhancement

Required heap to parse and emit CHIRRTL

| design | master (.fir) | this PR (.fir) | .pb on both |
| --- | --- | --- | --- |
| small | 800 M | 500 M | 300 M |
| medium | 9 G | 4.6 G | 3.4 G |
| large | 42 G | 13.5 G | 12 G |

Still not quite as good as `.pb` but a massive improvement. In profiling, the main objects being created in parsing are refs and subrefs, so perhaps we tweak the parsing rules again to not fully parse them and maybe parse them as part of a 2nd phase / during the CST => AST step?

### Contributor Checklist

- [None] Did you add Scaladoc to every public function/method?
- [NA] Did you update the FIRRTL spec to include every new feature/behavior?
- [ ] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement

<!-- Choose one or more from the following: -->

  - performance improvement   
  - new feature/API 

#### API Impact

Technically the generated Listener code (not the one I added, but the stuff generated by ANTLR) is public. I don't think it should be considered part of the FIRRTL public API but it's possible people would use it directly. We should probably move these files to package `firrtl.internal.antlr` or something to stop users from using them.

#### Backend Code Generation Impact

No impact

#### Desired Merge Strategy

 - Rebase

#### Release Notes

Improve performance and vastly reduce memory usage of `.fir` parser

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (1.2.x, 1.3.0, 1.4.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you mark as `Please Merge`?